requirements.txt
requests
google-generativeai
pillow
